requirements.txt
# Core dependencies for J1-J2 Heisenberg Prometheus Framework
# Pinned versions for reproducibility
# Last updated: 2024-02-14

# Configuration and validation
pydantic==2.5.3
pyyaml==6.0.1
typing-extensions==4.9.0

# Quantum physics and exact diagonalization
quspin==0.3.7
numpy==1.26.3
scipy==1.11.4

# Machine learning
torch==2.1.2
torchvision==0.16.2

# Data analysis and statistics
pandas==2.1.4
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Data storage
h5py==3.10.0

# Testing
pytest==7.4.4
hypothesis==6.92.2

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Additional dependencies
tqdm==4.66.1  # Progress bars
joblib==1.3.2  # Parallel processing utilities

# Jupyter notebook support (optional)
jupyter==1.0.0
ipykernel==6.28.0
notebook==7.0.6
